requirements.txt
# Ekran görüntüsü alma
mss>=9.0.1
Pillow>=10.0.0

# OCR (Optik Karakter Tanıma)
easyocr>=1.7.0
# Alternatif: pytesseract (Tesseract OCR gerekli)
# pytesseract>=0.3.10

# Görüntü işleme
opencv-python>=4.8.0
numpy>=1.24.0

# Klavye/Mouse kontrolü
pyautogui>=0.9.54
pynput>=1.7.6

# Yardımcı kütüphaneler
keyboard>=0.13.5
pywin32>=306
pyserial>=3.5

